+++ packages/examples/src/groovy/server/direct.ts
@@ -3,11 +3,11 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { resolve } from 'path';
-import { getLocalDirectory } from '../../utils/fs-utils.js';
+import { resolve } from 'node:path';
 import { runGroovyLanguageServer } from './main.js';
-const baseDir = resolve(getLocalDirectory(import.meta.url));
+import { getLocalDirectory } from '../../common/node/server-commons.js';
 
+const baseDir = resolve(getLocalDirectory(import.meta.url));
 const groovyJar = resolve(baseDir, '../../../resources/groovy/external/groovy-language-server-all.jar');
 const relativeDir = process.env.LANG_SERVER_JAR_PATH || groovyJar;
 console.log(`basedir: ${baseDir}`);