+++ packages/examples/src/json/server/main.ts
@@ -2,9 +2,10 @@
  * Copyright (c) 2024 TypeFox and others.
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
-import { runLanguageServer } from '../../common/language-server-runner.js';
-import { resolve } from 'path';
-import { LanguageName } from '../../common/server-commons.js';
+
+import { resolve } from 'node:path';
+import { runLanguageServer } from '../../common/node/language-server-runner.js';
+import { LanguageName } from '../../common/node/server-commons.js';
 
 export const runJsonServer = (baseDir: string, relativeDir: string) => {
     const processRunPath = resolve(baseDir, relativeDir);