+++ packages/examples/src/langium/langium-dsl/config/extendedConfig.ts
@@ -9,7 +9,7 @@ import '../../../../resources/vsix/GitHub.github-vscode-theme-6.3.4.vsix';
 import { useOpenEditorStub } from 'monaco-editor-wrapper/vscode/services';
 import { UserConfig } from 'monaco-editor-wrapper';
 import { loadLangiumWorker } from '../wrapperLangium.js';
-import { getTextContent } from '../../../utils/app-utils.js';
+import { getTextContent } from '../../../common/client/app-utils.js';
 
 export const setupLangiumClientExtended = async (): Promise<UserConfig> => {
     const code = await getTextContent(new URL('./src/langium/langium-dsl/content/example.langium', window.location.href));