+++ packages/examples/src/langium/statemachine/main.ts
@@ -8,7 +8,7 @@ import { createModelReference } from 'vscode/monaco';
 import { MonacoEditorLanguageClientWrapper } from 'monaco-editor-wrapper';
 import { useWorkerFactory } from 'monaco-editor-wrapper/workerFactory';
 import { createLangiumGlobalConfig } from './config/wrapperStatemachineConfig.js';
-import { getTextContent } from '../../utils/app-utils.js';
+import { getTextContent } from '../../common/client/app-utils.js';
 
 const wrapper = new MonacoEditorLanguageClientWrapper();
 const wrapper2 = new MonacoEditorLanguageClientWrapper();