+++ packages/examples/src/python/server/direct.ts
@@ -3,9 +3,9 @@
  * Licensed under the MIT License. See LICENSE in the package root for license information.
  * ------------------------------------------------------------------------------------------ */
 
-import { resolve } from 'path';
+import { resolve } from 'node:path';
 import { runPythonServer } from './main.js';
-import { getLocalDirectory } from '../../utils/fs-utils.js';
+import { getLocalDirectory } from '../../common/node/server-commons.js';
 
 const baseDir = resolve(getLocalDirectory(import.meta.url));
 const relativeDir = '../../../../../node_modules/pyright/dist/pyright-langserver.js';